move extensions out of allsettingfiles
divide up responsibilities so don't have to return meta information from allsettingfile creation

Fix bug where uploaded ext list was empty if ignored list was empty
